New password system
Remove the awful password system and replaced it with a popup window that you type your password into every time you rune the bot. This is a little tedious, but it's better than taking a plain text password from a file and putting it in 2 global variables. MCProtocolLib was also updated.